requirements.txt
docker
requests
pyyaml
python-dotenv
